requirements.txt
# YOLO Classifier Project Requirements
# ===================================
# Cross-platform dependencies for Windows, Linux, and macOS

# Core ML/AI Dependencies
torch>=2.0.0
torchvision>=0.15.0
yolov5>=7.0.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.7.0

# Data Handling
numpy>=1.21.0
pandas>=1.5.0
PyYAML>=6.0

# Data Augmentation (for training)
albumentations>=1.3.0

# GUI Framework (cross-platform)
# Note: tkinter is included with Python on most platforms

# Progress Bars and CLI
tqdm>=4.64.0

# HTTP Requests (for model downloads)
requests>=2.28.0

# Scientific Computing
scipy>=1.9.0
matplotlib>=3.5.0

# Development and Testing (optional)
pytest>=7.0.0; extra == "dev"
black>=22.0.0; extra == "dev"
flake8>=4.0.0; extra == "dev"

# Platform-specific optimizations
# These will be installed automatically based on the platform

# For better performance on supported platforms
psutil>=5.9.0

# Additional utilities
pathlib2>=2.3.0; python_version < "3.4" 